requirements.txt
grpcio-tools ~=1.60.0 
setuptools ~=67.7.2
wheel ~=0.40.0
spacy ~=3.5.2
grpcio-reflection ~=1.60.0